Changing flag names to human readable.
The existing Huddler thumbnail flag names are two letter forms such as
LS, LM, etc which aren't human readable or at least don't make sense out
of the Huddler context. To fix this, I changed them to the following:

	LS -> 'far' or fixed aspect ratio
	LL -> 'thumbnail-down'
	LM -> 'thumbnail'
	ZC -> 'zoom-crop'
	NS -> 'zoom-crop-down'

The 'far' option also has a $FILL component for the background color of
the delta between the aspect ratio of the image and the requested
dimensions. The default is set to white. We'll need to add this as an
optional parameter in the request. I'm thinking we'll probably want to
use a query parameter for this.