Refactor image-request-handler to a multimethod
We now need to dispatch on the :request-method which means adding an
exception for an unrecognized request method for robustness. It's not a
guideline, but doing this in the existing code would move the switch
statement to 3 cases including the else. Might as well use a multimethod
which will make it more clear. This also paves the way for adding
additional methods in the future (DELETE, HEAD, etc).