Move the height validation to bin/thumbnail
Move the height validation to the thumbnail script.

Previously we were setting {:height :auto} to pass the height
check in bin/thumbnail for all modes even though not all of them
required the height option. This is no longer necessary because we are
only performing that check for modes that require the height.